v1/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
diffusers>=0.10.0
transformers>=4.25.0
facenet-pytorch>=2.5.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0